+++ src/main/java/com/example/todo/api/TodoAction.java
@@ -2,14 +2,17 @@ package com.example.todo.api;
 
 import com.example.todo.application.TodoService;
 import com.example.todo.domain.*;
+
 import nablarch.core.repository.di.config.externalize.annotation.SystemRepositoryComponent;
 import nablarch.core.validation.ee.ValidatorUtil;
 import nablarch.fw.ExecutionContext;
+import nablarch.fw.jaxrs.EntityResponse;
 import nablarch.fw.web.HttpRequest;
 
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import nablarch.fw.web.HttpResponse;
 
 @SystemRepositoryComponent
 @Path("/todos/{todoId}")
@@ -24,25 +27,26 @@ public class TodoAction {
     @PUT
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public TodoResponse put(HttpRequest request, ExecutionContext context, PutRequest requestBody) {
+    public EntityResponse put(HttpRequest request, ExecutionContext context, PutRequest requestBody) {
         ValidatorUtil.validate(requestBody);
 
+        EntityResponse response = new EntityResponse();
         TodoId todoId = new TodoId(Long.valueOf(request.getParam("todoId")[0]));
         TodoStatus status = requestBody.completed ? TodoStatus.COMPLETED : TodoStatus.INCOMPLETE;
 
         Todo todo = todoService.updateStatus(todoId, status);
+        response.setEntity(new TodoResponse(todo.id(), todo.text(), todo.status()));
+        response.setStatusCode(HttpResponse.Status.CREATED.getStatusCode());
 
-        return new TodoResponse(todo.id(), todo.text(), todo.status());
+        return response;
+
     }
 
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
     public void delete(HttpRequest request, ExecutionContext context) {
-
         TodoId todoId = new TodoId(Long.valueOf(request.getParam("todoId")[0]));
-
         todoService.deleteTodo(todoId);
-
     }
 
     public static class PutRequest {
@@ -50,18 +54,4 @@ public class TodoAction {
         public Boolean completed;
     }
 
-    public static class TodoResponse {
-
-        public final Long id;
-
-        public final String text;
-
-        public final Boolean completed;
-
-        public TodoResponse(TodoId id, TodoText text, TodoStatus status) {
-            this.id = id.value();
-            this.text = text.value();
-            this.completed = status == TodoStatus.COMPLETED;
-        }
-    }
 }
